+++ IMXRT1060-EVK_APL/source/UserDriver/PCA9685/DrvPCA9685.h
@@ -68,7 +68,7 @@ typedef enum{
 }enPCA9685PortNo_t;
 
 extern _Bool DrvPCA9685Init(LPI2C_Type *base);
-extern _Bool DrvPCA9685SetPWMVal(LPI2C_Type *base, enPCA9685PortNo_t enPortNo, uint16_t PwmVal);
+extern _Bool DrvPCA9685SetPWMVal(LPI2C_Type *base, enPCA9685PortNo_t enPortNo, double PwmVal);
 extern _Bool DrvPCA9685GetPWMVal(LPI2C_Type *base, enPCA9685PortNo_t enPortNo, uint16_t *pPwmVal);
 
 #ifdef __cplusplus
